chore(leafletGroup) : extract leaflet extend class into it own class
